[test] Fix V1 (draft)
