+++ src/components/Common/Filters/RawFilter.vue
@@ -49,6 +49,7 @@
 <script>
 import JsonEditor from '../../Common/JsonEditor'
 import { formatFromBasicSearch } from '../../../services/filterManager'
+import { mapGetters } from 'vuex'
 
 export default {
   components: {
@@ -86,6 +87,7 @@ export default {
     }
   },
   computed: {
+    ...mapGetters('kuzzle', ['wrapper']),
     filterState() {
       try {
         return JSON.parse(this.rawFilter)
@@ -127,7 +129,7 @@ export default {
         }
         if (val.basic) {
           this.rawFilter = JSON.stringify(
-            this.formatFromBasicSearch(val.basic),
+            this.wrapper.basicSearchToESQuery(val.basic),
             null,
             2
           )
